feat: Add jar drawing functionality
Added the ability to draw a jar shape using cubic Bezier curves. Users can choose between a preset jar shape, a custom shape, or enter developer mode to see the individual sections of the jar. The program displays the jar shape as well as Bezier curves and allows saving the drawing as an image.